+++ src/frontend/components/CartIcon/CartIcon.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { CypressFields } from '../../utils/Cypress';
 import { useCart } from '../../providers/Cart.provider';
 import CartDropdown from '../CartDropdown';
 import * as S from './CartIcon.styled';
@@ -11,9 +12,9 @@ const CartIcon = () => {
 
   return (
     <>
-      <S.CartIcon onClick={() => setIsOpen(true)}>
+      <S.CartIcon data-cy={CypressFields.CartIcon} onClick={() => setIsOpen(true)}>
         <S.Icon src="/icons/Hipster_CartIcon.svg" alt="Cart icon" title="Cart" />
-        {!!items.length && <S.ItemsCount>{items.length}</S.ItemsCount>}
+        {!!items.length && <S.ItemsCount data-cy={CypressFields.CartItemCount}>{items.length}</S.ItemsCount>}
       </S.CartIcon>
       <CartDropdown productList={items} isOpen={isOpen} onClose={() => setIsOpen(false)} />
     </>